chore: have one client file per resource
