+++ client_chargepoints.go
@@ -2,6 +2,7 @@ package monta
 
 import (
 	"context"
+	"fmt"
 	"net/url"
 	"strconv"
 )
@@ -38,4 +39,10 @@ func (c *Client) ListChargePoints(
 		query.Set("teamId", strconv.Itoa(int(*request.TeamID)))
 	}
 	return doGet[ListChargePointsResponse](ctx, c, path, query)
+}
+
+// GetChargePoint to retrieve a single charge point.
+func (c *Client) GetChargePoint(ctx context.Context, chargePointID int64) (*ChargePoint, error) {
+	path := fmt.Sprintf("/v1/charge-points/%d", chargePointID)
+	return doGet[ChargePoint](ctx, c, path, nil)
 }
