--- client_charges_stop.go
@@ -1,12 +1,0 @@
-package monta
-
-import (
-	"context"
-	"fmt"
-)
-
-// StopCharge stop a charge.
-func (c *Client) StopCharge(ctx context.Context, chargeID int64) (*Charge, error) {
-	path := fmt.Sprintf("/v1/charges/%d/stop", chargeID)
-	return doPost[Charge](ctx, c, path, nil)
-}
